usecases decomposed
